Filter freezers by user ownership | graphEntities | json views | UnitTests passing(except logout)

+++ src/main/java/api/controller/MainController.java
@@ -1,10 +1,12 @@
 package api.controller;
 
+import java.util.Optional;
 import java.util.Set;
 
 import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.converter.json.MappingJacksonValue;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -12,10 +14,18 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import api.exceptionHandling.CustomException;
 import api.model.Aliment;
 import api.model.Freezer;
+import api.model.JsonViews;
+import api.model.User;
 import api.repository.AlimentRepository;
 import api.repository.FreezerRepository;
+import api.repository.UserRepository;
+import api.service.UserService;
+import io.vavr.collection.HashSet;
+
+// import io.vavr.collection.HashSet;
 
 @Controller
 @RequestMapping("/api")
@@ -26,6 +36,10 @@ public class MainController {
 	private FreezerRepository freezers;
 	@Autowired
 	private AlimentRepository aliments;
+	@Autowired
+	private UserRepository userRepo;
+	@Autowired
+	private UserService userService;
 
 	/* -------------------------------------------------------------------------- */
 	/*                                  freezers                                  */
@@ -33,23 +47,32 @@ public class MainController {
 	
 	@RequestMapping(path="/freezers", method=RequestMethod.POST, headers="Content-type=application/json")
 	@ResponseBody
-	public Freezer saveFreezer(HttpServletRequest request, @RequestBody Freezer freezer) {
+	public /*Freezer*/MappingJacksonValue saveFreezer(HttpServletRequest request, @RequestBody Freezer freezer) {
+		User currentUser = this.userService.getCurrentUser();		freezer.setUser(currentUser);
 		freezers.save(freezer);
-		return freezer;
+
+		final MappingJacksonValue result = new MappingJacksonValue(freezer);
+		result.setSerializationView(JsonViews.Summary.class);
+		return result;
 	}
 		
-	@RequestMapping(path="/freezers", method=RequestMethod.GET) //TODO: undo
+	@RequestMapping(path="/freezers", method=RequestMethod.GET)
 	@ResponseBody
-	public Set<Freezer> getFreezers(HttpServletRequest request) {
-		Set<Freezer> fromQuery = freezers.findAllWithAliments();
-
-		if ( request.getParameter("details") == null || ! request.getParameter("details").equals("complete")) {
-			for(Freezer item : fromQuery) {
-				item.setContent(null);
-			}
+	public /*Set<Freezer>*/MappingJacksonValue getFreezers(HttpServletRequest request) {
+		User currentUser = this.userService.getCurrentUser();
+		Class<?> jsonViewClass;
+		Set<Freezer> freezers;
+		if ( request.getParameter("details") == null || ! request.getParameter("details").equals("all")) {
+			freezers = userRepo.findByEmail_WithFreezers(currentUser.getEmail()).getFreezers();
+			jsonViewClass = JsonViews.Summary.class;
 		}
-
-		return fromQuery;
+		else {
+			freezers = userRepo.findByEmail_WithFreezersAndContent(currentUser.getEmail()).getFreezers();
+			jsonViewClass = JsonViews.Details.class;
+		}
+		final MappingJacksonValue result = new MappingJacksonValue(freezers);
+		result.setSerializationView(jsonViewClass);
+		return result;
 	}
 
 	@RequestMapping(path="/freezers/{idFreezer}", method=RequestMethod.PUT, headers="Content-type=application/json")
@@ -63,6 +86,11 @@ public class MainController {
 		sourceFreezer = freezers.findById(id).get();
 		if(sourceFreezer == null) {
 			throw new Error("Freezer could not be retrieved from the database.");
+		}
+
+		User currentUser = this.userService.getCurrentUser();
+		if(sourceFreezer.getUser() == null || sourceFreezer.getUser().getId() != currentUser.getId()) {
+			throw new Error("You cannot change a freezer you don't own.");
 		}
 
 		sourceFreezer.setName(freezer.getName());
@@ -79,10 +107,15 @@ public class MainController {
 			throw new Error("Invalid idFreezer in the url.");
 		}
 		Freezer sourceFreezer = null;
-		sourceFreezer = freezers.findByIdWithContent(id);
+		sourceFreezer = freezers.findByIdWithAliments(id).get();
 		
 		if(sourceFreezer == null) {
 			throw new Error("Freezer could not be retrieved from the database.");
+		}
+
+		User currentUser = this.userService.getCurrentUser();
+		if(sourceFreezer.getUser() == null || sourceFreezer.getUser().getId() != currentUser.getId()) {
+			throw new Error("You cannot delete a freezer you don't own.");
 		}
 
 		Set<Aliment> content = sourceFreezer.getContent();
@@ -100,15 +133,20 @@ public class MainController {
 	// @RequestMapping(path="/freezers/{idFreezer}/aliments", method=RequestMethod.GET, headers="Content-type=application/json")
 	@RequestMapping(path="/freezers/{idFreezer}/aliments", method=RequestMethod.GET)
 	@ResponseBody
-	public Set<Aliment> getFreezerContent(HttpServletRequest request, @PathVariable("idFreezer") String idFreezer) {		
-		return aliments.findFreezerContent(Long.valueOf(idFreezer));
+	public Set<Aliment> getFreezerContent(HttpServletRequest request, @PathVariable("idFreezer") String idFreezer) {
+		Optional<Freezer> freezer = freezers.findById(Long.valueOf(idFreezer));
+
+		checksIfFreezerOwned(freezer);
+		return freezer.get().getContent();
 	}
 	
 	@RequestMapping(path="/freezers/{idFreezer}/aliments", method=RequestMethod.POST, headers="Content-type=application/json")
 	@ResponseBody
 	public Aliment saveAliment(HttpServletRequest request, @PathVariable("idFreezer") String idFreezer, @RequestBody Aliment aliment) {
 
-		Freezer freezer = freezers.findById(Long.valueOf(idFreezer)).get();
+		Optional<Freezer> freezerOpt = freezers.findById(Long.valueOf(idFreezer));
+		checksIfFreezerOwned(freezerOpt);
+		Freezer freezer = freezerOpt.get();
 		//TODO:  catch NoSuchElementException for custom error output
 		aliment.setFreezer(
 			freezer
@@ -120,16 +158,21 @@ public class MainController {
 	
 	@RequestMapping(path="/freezers/{idFreezer}/aliments/{idAliment}", method=RequestMethod.PUT, headers="Content-type=application/json")
 	@ResponseBody
-	public Aliment editAliment(HttpServletRequest request, @PathVariable("idAliment") String idFreezer, @PathVariable("idAliment") String idAliment, @RequestBody Aliment aliment) {
+	public Aliment editAliment(HttpServletRequest request, @PathVariable("idFreezer") String idFreezer, @PathVariable("idAliment") String idAliment, @RequestBody Aliment aliment) {
+		
+		Optional<Freezer> freezerOpt = this.freezers.findByIdWithAliments(Long.valueOf(idFreezer));
+		checksIfFreezerOwned(freezerOpt);
+		Freezer freezer = freezerOpt.get();
 		Long id = Long.valueOf(idAliment);
 		if (id == null) {
 			throw new Error("Invalid idAliment in the url.");
 		}
-		Aliment sourceAliment = null;
-		try {
-			sourceAliment = aliments.findById(id).get();			
-		} catch(Exception e) {
-			System.out.println("error" + e.getMessage());
+		Aliment sourceAliment = HashSet.ofAll(freezer.getContent())
+			.find((Aliment alim) -> alim.getId() == id)
+			.get()
+		;
+		if(sourceAliment == null) {
+			throw new Error("The aliment could not be found in this freezer.");		
 		}
 		sourceAliment.setCategory(aliment.getCategory());
 		sourceAliment.setStoredDate(aliment.getStoredDate());
@@ -143,13 +186,33 @@ public class MainController {
 	
 	@RequestMapping(path="/freezers/{idFreezer}/aliments/{idAliment}", method=RequestMethod.DELETE)
 	@ResponseBody
-	public void deleteAliment(HttpServletRequest request, @PathVariable("idAliment") String idAliment) {
+	public void deleteAliment(HttpServletRequest request, @PathVariable("idFreezer") String idFreezer, @PathVariable("idAliment") String idAliment) {
+		Optional<Freezer> freezerOpt = freezers.findByIdWithAliments(Long.valueOf(idFreezer));
+		checksIfFreezerOwned(freezerOpt);
+
 		Long id = Long.valueOf(idAliment);
 		if (id == null) {
 			throw new Error("Invalid idAliment in the url.");
 		}
-		else {
-			aliments.deleteById(id);
+
+		Boolean isAlimentInFreezer = HashSet.ofAll(freezerOpt.get().getContent())
+			.exists((Aliment aliment) -> aliment.getId() == id);
+
+		if(!isAlimentInFreezer) {
+			throw new CustomException(String.format("The aliment of Id %s could not be found in freezer %s", idAliment, idFreezer));
 		}
+		
+		aliments.deleteById(id);
 	}
+
+	/* -------------------------------------------------------------------------- */
+	/*                                Utils Methods                               */
+	/* -------------------------------------------------------------------------- */
+	public void checksIfFreezerOwned(Optional<Freezer> freezer) {
+		if(!freezer.isPresent() || freezer.get().getUser() == null 
+			|| freezer.get().getUser().getId() != this.userService.getCurrentUser().getId()) {
+			throw new Error("You can only modify content of Freezers you own.");
+		}		
+	}
+
 }
